Add skip tutorial button + cleanup quest module
