+++ client/src/ui/components/quest/utils.tsx
@@ -1,0 +1,14 @@
+import { Quest } from "@/hooks/store/useQuestStore";
+
+export const groupQuestsByDepth = (quests: Quest[]): Record<number, Quest[]> => {
+  return quests?.reduce((groupedQuests: Record<number, Quest[]>, quest) => {
+    const depth = quest.depth;
+    if (!groupedQuests[depth]) {
+      groupedQuests[depth] = [];
+    }
+    groupedQuests[depth].push(quest);
+    return groupedQuests;
+  }, {});
+};
+
+export const areAllQuestsClaimed = (quests: Quest[]) => quests.every((quest) => quest.claimed);
